intern/chapter_001/src/test/java/ru/job4j/sbulygin/array/FindLoopTest.java
package ru.job4j.sbulygin.array;

import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * Class FindLoopTest.
 *
 * @author anonymous.
 * @version 1.0.
 * @since 27.01.19.
 */
public class FindLoopTest {
    /**
     * Test method indexOf for true value.
     */
    @Test
    public void whenArrayHasLength5Then0() {
        FindLoop find = new FindLoop();
        int[] input = new int[] {5, 10, 3};
        int value = 5;
        int result = find.indexOf(input, value);
        int expect = 0;
        assertThat(result, is(expect));
    }

    /**
     * Test method indexOf for false value.
     */
    @Test
    public void whenArrayHasLength15ThenMinusOne() {
        FindLoop find = new FindLoop();
        int[] input = new int[] {5, 10, 3};
        int value = 15;
        int result = find.indexOf(input, value);
        int expect = -1;
        assertThat(result, is(expect));
    }
}
